document pre-defined secret generators
There's a handful of secret generators already defined in
`agenix-rekey.nix`.  This brings them into awareness via some README
updates.  I did a cursory read on each of the various commands used and
believe most are accurate if not very close descriptions of what they
do, but please scrutinize this as some of them (like `dhparams`) I am
not very familiar with.