indicate which node is missing agenix and agenix-rekey modules
When running `agenix rekey`, if a node (typically a host) doesn't have both
`agenix` and `agenix-rekey` included as modules, you will see an error like
this:

```
       error: attribute 'age' missing
       at /nix/store/zy9vgszqqqh51wmk8cjj6hghx6fydwbi-source/apps/rekey.nix:41:43:
           40|
           41|   nodesWithDerivationStorage = filter (x: x.config.age.rekey.storageMode == "derivation") (attrValues nodes);
             |                                           ^
           42|
```

This error doesn't tell you which host is missing the modules, let alone that
the modules should be included.  The `agenix rekey` tool has this information
available and can provide it upon encountering this issue.  These changes will
detect the missing keys and halt with an error indicating which node is causing
the problem and that the required modules (`agenix` and `agenix-rekey`) must be
included.  The user will still encounter the error in the same place and there
is no change in actual functionality.  Now this is what they will see:

```
       error: Node "my-host" is missing the agenix module.
       agenix-rekey cannot continue until all nodes include the agenix module.
```

When only `agenix-rekey` is missing:

```
       error: Node "my-host" is missing the agenix-rekey module.
       agenix-rekey cannot continue until all nodes include the agenix-rekey module.
```