do not add newlines in generated secrets
`<<< "$var"` is the equivalent to `printf '%s\n' "$var"`.  This means we add an
extra newline to our secrets as we encrypt them.  Many systems inspect their
secret files literally, and interpret the newline as just another character in
the secret (as opposed to a POSIX end-of-line+file).  This changes our
encryption to use a `printf` expression that does not add a newline, and thus
means we no longer mangle secrets emitted to disk.

This should have no impact on systems that filtered the newline in the first
place.