feat: default parent dir for non-generated secrets
that permits specifying non-generated secret paths in relative form.

BREAKING CHANGE: brings the logic of `generatedSecretsDir` in line with
its option description ("sets `age.secrets.<name>.rekeyFile` to a
default value in this directory, for any secret that defines a
generator"), whereas previously *all* secrets were expanded relative to
`generatedSecretsDir` if it was defined.